+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/messages/MessageListViewModel.kt
@@ -49,6 +49,7 @@ import io.getstream.chat.android.offline.ChatDomain
 import io.getstream.chat.android.offline.channel.ChannelController
 import io.getstream.chat.android.offline.experimental.channel.thread.state.ThreadState
 import io.getstream.chat.android.offline.experimental.extensions.asReferenced
+import io.getstream.chat.android.offline.extensions.loadOlderMessages
 import io.getstream.chat.android.offline.model.ConnectionState
 import io.getstream.chat.android.offline.thread.ThreadController
 import kotlinx.coroutines.Job
@@ -439,7 +440,7 @@ public class MessageListViewModel(
                 .enqueue()
         } else {
             messagesState = messagesState.copy(isLoadingMore = true)
-            chatDomain.loadOlderMessages(channelId, messageLimit).enqueue()
+            chatClient.loadOlderMessages(channelId, messageLimit).enqueue()
         }
     }
 
@@ -520,7 +521,7 @@ public class MessageListViewModel(
                 messageActions = messageActions + messageAction
             }
             is Copy -> copyMessage(messageAction.message)
-            is MuteUser -> muteUser(messageAction.message.user)
+            is MuteUser -> updateUserMute(messageAction.message.user)
             is React -> reactToMessage(messageAction.reaction, messageAction.message)
             is Pin -> updateMessagePin(messageAction.message)
             else -> {
@@ -714,12 +715,18 @@ public class MessageListViewModel(
     }
 
     /**
-     * Mutes the user that sent a particular message.
+     * Mutes or unmutes the user that sent a particular message.
      *
-     * @param user The user to mute.
+     * @param user The user to mute or unmute.
      */
-    private fun muteUser(user: User) {
-        chatClient.muteUser(user.id).enqueue()
+    private fun updateUserMute(user: User) {
+        val isUserMuted = chatDomain.muted.value.any { it.target.id == user.id }
+
+        if (isUserMuted) {
+            chatClient.unmuteUser(user.id)
+        } else {
+            chatClient.muteUser(user.id)
+        }.enqueue()
     }
 
     /**