+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/querychannels/QueryChannelsControllerIntegrationTest.kt
@@ -5,9 +5,6 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
-import com.nhaarman.mockitokotlin2.reset
-import com.nhaarman.mockitokotlin2.verify
-import com.nhaarman.mockitokotlin2.verifyZeroInteractions
 import com.nhaarman.mockitokotlin2.whenever
 import io.getstream.chat.android.client.api.models.FilterObject
 import io.getstream.chat.android.client.api.models.QuerySort
@@ -158,50 +155,6 @@ internal class QueryChannelsControllerIntegrationTest : BaseConnectedMockedTest(
             )
 
             sut.channels.value shouldBeEqualTo listOf(data.channel1, data.channel2, data.channel3)
-        }
-
-    @Test
-    fun `Given channel in the list and default events handler When handle NotificationMessageNewEvent Should not make request to API`(): Unit =
-        runBlocking {
-            testCoroutines.scope.launch {
-                val sut = Fixture(chatDomainImpl, data.filter1)
-                    .givenChannelsInOfflineStorage(data.channel1, data.channel2)
-                    .givenChannelEventsHandler(DefaultChatEventHandler(client, queryControllerImpl.channels))
-                    .get()
-                sut.query()
-                reset(client)
-
-                sut.handleEvent(
-                    mock<NotificationMessageNewEvent> {
-                        on { it.channel } doReturn data.channel2
-                        on { it.cid } doReturn data.channel2.cid
-                    }
-                )
-
-                verifyZeroInteractions(client)
-            }.join()
-        }
-
-    @Test
-    fun `Given channel is not in the list and default events handler When handle NotificationMessageNewEvent Should make request to API`(): Unit =
-        runBlocking {
-            val sut = Fixture(chatDomainImpl, data.filter1)
-                .givenChannelsInOfflineStorage(data.channel1)
-                .givenChannelEventsHandler(DefaultChatEventHandler(client, queryControllerImpl.channels))
-                .get()
-            sut.query()
-            reset(client)
-            whenever(client.channel(any())) doReturn channelClientMock
-            whenever(client.queryChannelsInternal(any())) doReturn emptyList<Channel>().asCall()
-
-            sut.handleEvent(
-                mock<NotificationMessageNewEvent> {
-                    on { it.channel } doReturn data.channel2
-                    on { it.cid } doReturn data.channel2.cid
-                }
-            )
-
-            verify(client).queryChannelsInternal(any())
         }
 
     private class Fixture(